feat(parser): created markdown parser
This initial version of the markdown parser uses regular expressions to convert an input string into an output string (in html). This is not the best solution but was a quick way to implement a parser for the initial version of the project.

This parser will most likely be replaced with a more robost solution in the future.